Enable persistence queue (PQ) and reduce the number of logstash instances to 2, each one publishing its own ports.
Persistence Queue (PQ) helps our logstash service to avoid data loss at the minor cost of performance. It uses disk instead of memory and ensures that in case of sudden shutdown of the service, the in-flight events will still make their way into our elasticsearch database.

Alos, as was recommended with the documentation, I only kept 2 instances of logstash for high availability and decided to use elasticsearch inputs' built-in load balancer instead of NginX. I'm not sure if that helps with the performance but I guess the built-ins are always better.